feat(architecture): separate weather service from subscription
