+++ services/notification/.dockerignore
@@ -1,0 +1,5 @@
+node_modules
+dist
+*.log
+*.md
+.git
