--- services/subscription/src/common/cache/cache.service.ts
@@ -1,8 +1,0 @@
-export abstract class CacheService {
-  abstract get<T = any>(key: string): Promise<T | null>;
-  abstract set<T = any>(
-    key: string,
-    value: T,
-    ttlSeconds: number,
-  ): Promise<void>;
-}
