+++ services/subscription/src/common/constants/http.constants.ts
@@ -1,9 +1,4 @@
 export const HTTP_ERROR_MESSAGES = {
-  // Weather
-  WEATHER_INVALID_REQUEST: 'Invalid request',
-  WEATHER_CITY_NOT_FOUND: 'City not found',
-
-  // Subscription
   INVALID_INPUT: 'Invalid input',
   EMAIL_ALREADY_SUBSCRIBED: 'Email already subscribed',
   INVALID_TOKEN: 'Invalid token',
