+++ services/subscription/src/common/types/weather-data.type.ts
