+++ services/subscription/src/config/app-config.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-import { WeatherProviderConfig } from 'src/weather/domain/types/weather-provider-config.type';
 import { ConfigService } from '@nestjs/config';
 
 @Injectable()
@@ -23,22 +22,6 @@ export class AppConfigService {
       migrationsRun: true,
       migrations: ['dist/database/migrations/*.js'],
       logging: ['query', 'error', 'schema', 'warn'],
-    };
-  }
-
-  getOpenWeatherMapConfig(): WeatherProviderConfig {
-    return {
-      baseUrl: this.configService.getOrThrow<string>(
-        'OPENWEATHER_API_BASE_URL',
-      ),
-      apiKey: this.configService.getOrThrow<string>('OPENWEATHER_API_KEY'),
-    };
-  }
-
-  getWeatherApiConfig(): WeatherProviderConfig {
-    return {
-      baseUrl: this.configService.getOrThrow<string>('WEATHER_API_BASE_URL'),
-      apiKey: this.configService.getOrThrow<string>('WEATHER_API_KEY'),
     };
   }
 }
