+++ services/subscription/src/config/validation.ts
@@ -6,14 +6,4 @@ export const validationSchema = Joi.object({
   POSTGRES_USER: Joi.string().required(),
   POSTGRES_PASSWORD: Joi.string().required(),
   POSTGRES_DB: Joi.string().required(),
-
-  WEATHER_API_BASE_URL: Joi.string().uri().required(),
-  WEATHER_API_KEY: Joi.string().required(),
-
-  OPENWEATHER_API_BASE_URL: Joi.string().uri().required(),
-  OPENWEATHER_API_KEY: Joi.string().required(),
-
-  REDIS_HOST: Joi.string().required(),
-  REDIS_PORT: Joi.string().pattern(/^\d+$/).required(),
-  REDIS_TTL: Joi.string().pattern(/^\d+$/).required(),
 });
