--- services/subscription/src/monitoring/domain/metrics.service.ts
@@ -1,4 +1,0 @@
-export abstract class MetricsService {
-  abstract incWeatherCacheHit(): void;
-  abstract incWeatherCacheMiss(): void;
-}
