+++ services/subscription/src/subscription/infrastructure/notification/dto/confirmation-email.dto.ts
