+++ services/subscription/src/subscription/subscription.module.ts
@@ -5,18 +5,17 @@ import { SubscriptionService } from './application/subscription.service';
 import { SubscriptionController } from './interfaces/controllers/subscription.controller';
 import { SubscriptionEntity } from './infrastructure/persistence/entities/subscription.entity';
 import { TypeOrmSubscriptionRepository } from './infrastructure/persistence/repositories/typeorm-subscription.repository';
-import { WeatherModule } from 'src/weather/weather.module';
 import { SubscriptionCronService } from './application/cron/subscription.cron';
 import { ScheduleModule } from '@nestjs/schedule';
 import { TokenModule } from 'src/token/token.module';
-import { NotificationHttpService } from './application/notification/notification-http-service';
+import { NotificationHttpService } from './infrastructure/notification/notification-http-service';
 import { HttpModule } from '@nestjs/axios';
+import { WeatherHttpClientService } from './infrastructure/weather/weather-http.client';
 @Module({
   imports: [
     TypeOrmModule.forFeature([SubscriptionEntity]),
     ScheduleModule.forRoot(),
     ConfigModule,
-    WeatherModule,
     TokenModule,
     HttpModule,
   ],
@@ -25,6 +24,7 @@ import { HttpModule } from '@nestjs/axios';
     SubscriptionService,
     SubscriptionCronService,
     NotificationHttpService,
+    WeatherHttpClientService,
     {
       provide: 'SubscriptionRepository',
       useClass: TypeOrmSubscriptionRepository,
