--- services/subscription/src/weather/domain/interfaces/weather-provider.interface.ts
@@ -1,5 +1,0 @@
-import { WeatherData } from '../types/weather-data.type';
-
-export interface WeatherProvider {
-  getWeather(city: string): Promise<WeatherData>;
-}
