--- services/subscription/src/weather/domain/types/weather-provider-config.type.ts
@@ -1,4 +1,0 @@
-export type WeatherProviderConfig = {
-  baseUrl: string;
-  apiKey: string;
-};
