--- services/subscription/src/weather/domain/weather.model.ts
@@ -1,7 +1,0 @@
-export class Weather {
-  constructor(
-    public readonly temperature: number,
-    public readonly humidity: number,
-    public readonly description: string,
-  ) {}
-}
