--- services/subscription/src/weather/infrastructure/logger/weather-logger.config.ts
@@ -1,7 +1,0 @@
-import * as path from 'path';
-
-export const WEATHER_LOG_FILE_PATH = path.resolve(
-  process.cwd(),
-  'logs',
-  'weather.log',
-);
