--- services/subscription/src/weather/infrastructure/providers/types/openweathermap-response.type.ts
@@ -1,9 +1,0 @@
-export type OpenWeatherMapResponse = {
-  main: {
-    temp: number;
-    humidity: number;
-  };
-  weather: {
-    description: string;
-  }[];
-};
