--- services/subscription/src/weather/infrastructure/providers/types/weatherapi-response.type.ts
@@ -1,12 +1,0 @@
-export type WeatherApiResponse = {
-  location: {
-    name: string;
-  };
-  current: {
-    temp_c: number;
-    humidity: number;
-    condition: {
-      text: string;
-    };
-  };
-};
